+++ src/main/java/personal/bookerav2/dto/reviews/ReviewBookDto.java
@@ -7,6 +7,6 @@ public record ReviewBookDto(
         String username,
         String avatarUrl,
         String content,
-        Integer rating
+        Short rating
 ) {
 }
